Add files via upload
Removes transitions from the data and exports the resulting data as a csv. 

My reccomendation is to use the boolean version. It is faster. The numeric one might be something we mess around with if we need the transition types (MPI, MPO, BSI and BSO) but for the moment isn't needed.